Making types consistent and fixing comments.

+++ Project-2/Problem-4/p2-4.cpp
@@ -8,7 +8,7 @@
 #include <iomanip>
 
 // TODO: Move this declaration to a h-file.
-void jacobi_rotate(arma::mat &A, arma::mat &R, int k, int l);
+void jacobi_rotate(arma::mat &A, arma::mat &R, size_t k, size_t l);
 
 int main()
 {
